sidebar bottom component; remove tabs (chats/tasks)
- remove tabs, unify chats/tasks
- sidebar bottom section component
- guards for speechStore and sidebarStore for double initialization
