+++ webui/components/chat/input/bottom-actions.html
@@ -1,7 +1,13 @@
 <html>
-<head></head>
+<head>
+  <script type="module">
+    import { store } from "/components/chat/input/input-store.js";
+  </script>
+</head>
 <body>
-    <div class="text-buttons-row" x-data>
+  <div x-data>
+    <template x-if="$store.chatInput">
+      <div class="text-buttons-row">
         <button class="text-button" @click="$store.chatInput.pauseAgent(!$store.chatInput.paused)">
             <template x-if="!$store.chatInput.paused">
                 <svg xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" width="14" height="14">
@@ -54,9 +60,17 @@
             </svg>
             <p>Nudge</p>
         </button>
-    </div>
+      </div>
+    </template>
+  </div>
 
-    <style>
+  <script>
+    document.addEventListener('alpine:init', () => {
+      const s = Alpine.store('chatInput');
+      if (s && typeof s.init === 'function') s.init();
+    });
+  </script>
+  <style>
         .text-buttons-row { width: 100%; display: flex; padding-top: var(--spacing-xs); margin-left: var(--spacing-xs); }
         .text-button { background-color: transparent; border: none; border-radius: 5px; color: var(--color-text); font-family: "Rubik", Arial, Helvetica, sans-serif; font-size: 0.6rem; padding: 6px var(--spacing-sm); cursor: pointer; opacity: 0.8; -webkit-transition: all 0.3s; transition: all 0.3s; display: flex; align-items: center; gap: var(--spacing-xs); }
         .text-button:hover { opacity: 1; background-color: var(--color-secondary); border-radius: 4px; }
