+++ webui/components/sidebar/chats/chats-list.html
@@ -8,7 +8,8 @@
   <div x-data>
     <template x-if="$store.chats">
       <div class="chats-list-container" x-data>
-        <ul class="config-list" x-show="$store.chats.contexts.length > 0">
+        <h3 class="section-header">Chats</h3>
+        <ul class="config-list chats-config-list" x-show="$store.chats.contexts.length > 0">
           <template x-for="context in $store.chats.contexts" :key="context.id">
             <li>
               <div :class="{'chat-list-button': true, 'font-bold': context.id === $store.chats.selected}"
@@ -39,22 +40,34 @@
 <style>
   /* Chats list container and items */
   .chats-list-container {
+    display: flex;
+    flex-direction: column;
+    flex: 1;
+    min-height: 0;
+    overflow: hidden;
+  }
+
+  /* Sticky header for chats list */
+  .chats-list-container .section-header {
+    flex-shrink: 0;
+    z-index: 10;
+  }
+
+  /* Scrollable config list */
+  .chats-config-list {
     flex: 1;
     min-height: 0;
     overflow-y: auto;
     scroll-behavior: smooth;
-    /* Fade mask */
-    mask-image: linear-gradient(to bottom, black calc(100% - 20px), transparent 100%);
-    -webkit-mask-image: linear-gradient(to bottom, black calc(100% - 20px), transparent 100%);
-    background: linear-gradient(to bottom, calc(100% - 20px), transparent 100%);
-    padding-bottom: 20px;
-    scrollbar-width: thin;
-    -ms-overflow-style: auto;
+    max-height: 23vh;
+    /* Hide scrollbar but keep functionality */
+    scrollbar-width: none; /* Firefox */
+    -ms-overflow-style: none; /* IE/Edge */
   }
-  .chats-list-container::-webkit-scrollbar { width: 5px; }
-  .chats-list-container::-webkit-scrollbar-track { background: rgba(0, 0, 0, 0.2); border-radius: 6px; }
-  .chats-list-container::-webkit-scrollbar-thumb { background-color: var(--color-border); border-radius: 6px; }
-  .chats-list-container::-webkit-scrollbar-thumb:hover { background-color: var(--color-border); }
+  /* Hide scrollbar for Chrome/Safari/Webkit */
+  .chats-config-list::-webkit-scrollbar {
+    display: none;
+  }
 
   .chat-list-button { display: block; width: 100%; padding: 8px 5px; cursor: pointer; overflow: hidden; position: relative; border-radius: 4px; transition: background-color 0.2s ease-in-out; }
   .chat-list-button.has-task-container { padding-top: 6px; padding-bottom: 6px; }
