+++ webui/components/sidebar/sidebar-store.js
@@ -3,27 +3,17 @@ import { createStore } from "/js/AlpineStore.js";
 // This store manages the visibility and state of the main sidebar panel.
 const model = {
   isOpen: true,
-  versionNo: "",
-  commitTime: "",
-
-  get versionLabel() {
-    return this.versionNo && this.commitTime
-      ? `Version ${this.versionNo} ${this.commitTime}`
-      : "";
-  },
+  _initialized: false,
 
   // Initialize the store by setting up a resize listener
+  // Guard ensures this runs only once, even if called from multiple components
   init() {
+    if (this._initialized) return;
+    this._initialized = true;
+
     this.handleResize();
     this.resizeHandler = () => this.handleResize();
     window.addEventListener("resize", this.resizeHandler);
-    
-    // Load version info from global scope
-    const gi = (globalThis && globalThis.gitinfo) ? globalThis.gitinfo : null;
-    if (gi && gi.version && gi.commit_time) {
-      this.versionNo = gi.version;
-      this.commitTime = gi.commit_time;
-    }
   },
 
   // Cleanup method for lifecycle management
@@ -32,6 +22,7 @@ const model = {
       window.removeEventListener("resize", this.resizeHandler);
       this.resizeHandler = null;
     }
+    this._initialized = false;
   },
 
   // Toggle the sidebar's visibility
