+++ webui/index.html
@@ -134,7 +134,6 @@
         import { store as sidebarStore } from "/components/sidebar/sidebar-store.js";
         import { store as chatsStore } from "/components/sidebar/chats/chats-store.js";
         import { store as tasksStore } from "/components/sidebar/tasks/tasks-store.js";
-        import { store as tabsStore } from "/components/sidebar/tabs-store.js";
     </script>
     <script>
         // Expose git info for sidebar component
